requirements.txt
langchain>=0.2.0
llama-index>=0.10.0
openai>=1.6.0
anthropic>=0.17.0
groq>=0.4.0
chromadb>=0.4.22
pinecone-client>=3.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0
gradio>=3.47.1
pandas>=2.0.0
numpy>=1.25.0
tiktoken>=0.5.2
httpx>=0.25.0,<0.28.0  # pin below 0.28 because OpenAI client passes proxies kw
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
slowapi>=0.1.9
python-multipart>=0.0.6
reportlab>=4.0.0
